refactor: clean up code alot
