+++ community-stabilization-fund-api/pages/admin/users/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 
-import UserService from '../../../src/services/users';
+import UserService from '../../../src/services/user';
 
 import type { User } from '../../../src/db';
 import type { NextPage } from 'next';
