+++ community-stabilization-fund-api/pages/api/users/[id].ts
@@ -6,7 +6,8 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 const userHandler = (req: NextApiRequest, res: NextApiResponse) => {
   const { method, query, body } = req;
   const userId = query.id as string;
-
+  console.log({ userId, method, query, body});
+  
   switch (method) {
     case 'GET':
       getUserById(userId, res);
@@ -26,44 +27,58 @@ const userHandler = (req: NextApiRequest, res: NextApiResponse) => {
 
 const getUserById = async (id: string, res: NextApiResponse) => {
   const sql = queries.makeGetByIdSql('api_user');
-  const user: User = await executeQuery({ sql, values: [id] });
 
-  if (!user) {
-    return res.status(404).json({
-      status: 404,
-      message: 'Not Found',
-    });
+  try {
+    const user: User = await executeQuery({ sql, values: [id] });
+
+    if (!user) {
+      return res.status(404).json({
+        status: 404,
+        message: 'Not Found',
+      });
+    }
+
+    return res.json({ ...user });
   }
+  catch (error) { 
 
-  return res.json({ ...user });
+  }
 };
 
 const updateUserById = async (body: any, res: NextApiResponse) => {
   const sql = queries.makeUpdateSql('api_user', body, `id=${body.id}`);
 
   try {
-    const result = await executeQuery({
-      sql,
-    });
-
+    const result = await executeQuery({sql});
+    if(!result) {
+      return res.status(404).json({
+        status: 404,
+        message: 'Not Found',
+      });
+    }
+  
     return res.json(result);
   } catch (error) {
-    console.log(error);
+    return res.json({error});
   }
 };
 
 const deleteUserById = async (id: string, res: NextApiResponse) => {
   const sql = queries.makeDeleteSql('api_user');
-  const results = await executeQuery({ sql, values: [id] });
 
-  if (!results) {
-    return res.status(404).json({
-      status: 404,
-      message: 'Not Found',
-    });
+  try {
+    const results = await executeQuery({ sql, values: [id] });
+    if (!results) {
+      return res.status(404).json({
+        status: 404,
+        message: 'Not Found',
+      });
+    }
+  
+    return res.send('Successfully deleted user with id: ' + id);
+  } catch (error) {
+    return res.json({error});
   }
-
-  return res.send('Successfully deleted user with id: ' + id);
 };
 
 export default userHandler;
