+++ community-stabilization-fund-api/pages/api/users/authenticate/index.tsx
@@ -2,7 +2,7 @@ import { executeQuery, queries } from "../../../../src/db";
 
 import type { NextApiRequest, NextApiResponse } from "next";
 
-const userHandler = (req: NextApiRequest, res: NextApiResponse) => {
+const authHandler = (req: NextApiRequest, res: NextApiResponse) => {
   const { method, body } = req;
 
   switch (method) {
@@ -20,15 +20,19 @@ const authenticateUser = async (body: any, res: NextApiResponse) => {
   const { apiUser, token } = body; //foo:bar
   const sql = queries.makeAuthenticateSql(apiUser, token);
 
-  const result = await executeQuery({ sql });
-  const { error } = result;
-
-  if(error) return res.status(400).send({ error });
-  if(result.length === 0) return res.status(401).send(`ERROR: There is no apikey matching ${apiUser}:${token}`);
-
-  const id = result[0]?.id;
-  return res.status(200).json({ id });
+  try {
+    const result = await executeQuery({ sql });
+  
+    if(result.length === 0) {
+      return res.status(401).send(`ERROR: There is no apikey matching ${apiUser}:${token}`);
+    }
+  
+    const id = result[0]?.id;
+    return res.status(200).json({ id });
+  } catch (error) {
+    return res.status(400).send({ error });
+  }
 };
 
 
-export default userHandler;
+export default authHandler;
