+++ community-stabilization-fund-api/src/services/index.ts
@@ -1,1 +1,1 @@
-export * from './users';
+export * from './user';
